devapp/owners: add owners for golang.org/x/tools
This change is part of an effort to improve the maintenance of the
x/tools repository. Historically, we have not had well-defined owners
for tools, which has caused a number of CLs to slip through the cracks.

I've added as many owners here as I know or can guess. Please correct me
if I'm mistaken. For any package that doesn't seem to have owners or
that I wasn't sure about, I've just listed "defaultToolsOwners" as a
placeholder so that we don't forget about those packages.

I've added everyone that is a part of this change as a reviewer to
make sure they are aware of it. Please +1 to indicate that you
are OK with the changes.

Change-Id: I6a2d7bb48d2c72910a1407ce6f10cc1c03288d77
Reviewed-on: https://go-review.googlesource.com/c/build/+/228217
Run-TryBot: Rebecca Stambler <rstambler@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>