cmd/fetchlogs: fetch repo data using the 'repo' query param
The repo flag for fetchlogs is not currently that useful because it
queries the main Go dashboard data, which only has a few commits for
each subrepo. This was probably because the data at
  https://build.golang.org/?repo=<import path>&mode=json
was incorrect due to golang/go#35515.

With golang.org/cl/232897, this data should be corrected. Update
fetchlogs to now query data from the subrepo dashboard.

Change-Id: I352662abf7da6abb7bc23888b11e03927f567cab
Reviewed-on: https://go-review.googlesource.com/c/build/+/232898
Run-TryBot: Robert Findley <rfindley@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>