cmd/releasebot: use same master commit for beta releases
The prepare mode is responsible for doing testing to ensure
all tests are passing. If a problem is found, it can be fixed
and the prepare mode should be re-done.

The release mode is responsible for taking the well-tested
commit from the prepare mode, and using it to build the release.

For beta releases, the master branch is used, and a commit has a
chance of landing between the time the prepare and release modes
are started.

Instead of fetching the latest master commit in release mode,
start reusing the same master commit that was already tested
in prepare mode.

Fixes golang/go#36182.
For golang/go#36171.

Change-Id: Ibf8052ef385cb0a1ffed372ae3e0ed0a9575882f
Reviewed-on: https://go-review.googlesource.com/c/build/+/235777
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>