app/appengine: stop hiding some tested configurations for x repos
The build dashboard can be used to view builds for the main Go repo
and other golang.org/x repos. The isUntested invocation was invalid
for repos other than the main one, which was causing results for some
tested configurations to become hidden (incorrectly replaced by '•').

Larger changes are needed before there's sufficient data to compute
the GoBranch value for all repos reliably, so for now, just update
the isUntested invocation to apply only for the main repo.

This prioritizes the ability to view test results over the ability
to see that some builds are intentionally missing because they are
configured not to run. Doing both is a part of future work.

For golang/go#40290.
For golang/go#34744.
For golang/go#28643.

Change-Id: Id43cb47abacb1036f578efbb8232ae17ad40eca9
Reviewed-on: https://go-review.googlesource.com/c/build/+/244137
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>