dashboard: create misc-compile-ios builder
Create the builder as a post-submit builder with a non-zero KnownIssue
value first. Once we know it works well, we'll have the confidence to
be able to safely promote it to a normal TryBot, without the risk of
causing disruption for contributors.

For golang/go#41610.

Change-Id: If871addf325b7728109afbe094b72d40e27ceec1
Reviewed-on: https://go-review.googlesource.com/c/build/+/257617
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>