dashboard: create misc-compile-darwinarm64 builder
Create the builder as a post-submit builder with a non-zero KnownIssue
value first. Once we know it works well, we'll have the confidence to
be able to safely promote it to a normal TryBot, without the risk of
causing disruption for contributors.

It's different from misc-compile-darwin since it needs to run on recent
Go 1.16 commits only.

For golang/go#42341.

Change-Id: I67170988d7323cb73439755a3961532c668040e7
Reviewed-on: https://go-review.googlesource.com/c/build/+/267105
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Cherry Zhang <cherryyz@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>