cmd/releasebot, cmd/release: add 2 more longtest test-only targets
Add support for a linux-386-longtest test-only target. This is the
third longtest builder that we have.

Both linux-{386,amd64}-longtest targets are passing on master
for future Go 1.17 releases, as well as release-branch.go1.16
and release-branch.go1.15 now that CL 288818 is cherry-picked
(see backport issues golang/go#45240 and golang/go#45239).

Start including them in the release process by default, so that we
have more automated test coverage and more information available
during the upcoming releases, and in turn can use it to improve
the releases and the release process further.

Fixes golang/go#29252.

Change-Id: I9740c9ce5cae73a3ed9fe320219f1044f5fdcc9a
Reviewed-on: https://go-review.googlesource.com/c/build/+/304949
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>