cmd/racebuild: add darwin/arm64 support
In Go 1.16 we added darwin/arm64 (macOS) port, along with race
detector support. At that time we didn't have a builder, so the
syso file was built manually. Now that we have a builder, this CL
adds darwin/arm64 support to racebuild, so the syso file can be
built with it.

Verified that it successfully builds the syso and it passes the
tests.

Change-Id: I0bac1345a0b5f1df74dcb47cec13b6856a647094
Reviewed-on: https://go-review.googlesource.com/c/build/+/318110
Trust: Cherry Mui <cherryyz@google.com>
Run-TryBot: Cherry Mui <cherryyz@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>