cmd/buildlet: support Windows ARM64 qemu guests
The Windows ARM builders on EC2 run on qemu via KVM. We use a userspace
network device for simplicity, which also conveniently restricts which
ports are exposed. In order to make a request to the metatadata service,
which is routed in a different way than the public internet, we
explicitly forward the port to the guest VM on a special ip/port.

This introduces a hard-coded value for Windows ARM64 buildlets. We
should be able to improve this code to detect it, but this will get the
buildlet unblocked.

The buildlet image also includes llvm-arm64 mingw rather than a gcc
based distribution. This change also adds the correct directory to the
path.

For golang/go#42604

Change-Id: Ife2ebb900a08034d6e0dfa0982a24b312ee6d70a
Reviewed-on: https://go-review.googlesource.com/c/build/+/322653
Trust: Alexander Rakoczy <alex@golang.org>
Run-TryBot: Alexander Rakoczy <alex@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>