cmd/coordinator: link "ok" results to build logs
Every now and then, it's highly desirable to be able to view build
logs for a build even if it was successful. Those logs are kept as
of CL 198197, but they just haven't been linked. At long last, add
those links to the build dashboard(s)! 🎉

Keep the visual style of "ok" cells the same, aside that the mouse
cursor becomes a hand with an index finger, rather than a text caret.
This means this feature isn't very discoverable, but we can address
that in the future. (I've experimented with giving the "ok" links
an underline, but even a faint underline quickly leads to a large
amount of overwhelming visual noise when there are hundreds of "ok"
cells on many pages.)

Apply the same change to the work-in-progress dashboard V2 page,
to keep it in sync. (There are additional CSS changes to make it
look more consistent with the current build.golang.org dashboard.)

For golang/go#34119.
Updates golang/go#34744.

Change-Id: I27530e4103bbfd10e8cf09ae1a1cfce06c9bd748
Reviewed-on: https://go-review.googlesource.com/c/build/+/340435
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>