devapp: automatically deduce current milestone for burndown chart
Use information from maintner to figure out the current development
milestone, and estimate the start of work based on the current six-
month cycle. It should work well enough for now, and can be changed
as needed.

This way it's no longer necessary to manually update the current
milestone and its date every 6 months, as was previously done in
CLs 295191, 248157, 228110, and 193877. Hooray for gophers doing
more of the work for us!

Change-Id: I10b569da577e638a190886c7678f6da1107fdb46
Reviewed-on: https://go-review.googlesource.com/c/build/+/344089
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>