devapp/owners: add mvdan as secondary to cmd/gofmt
I've contributed sporadically over the years,
and more recently by experimenting externally in gofumpt.

With a CL in-flight to make gofmt format files in parallel,
it feels like it's time to make this a bit more official.

The main important bits are cmd/gofmt, the tool,
and go/printer, the guts of the tool.
go/format is pulled in as a library version of cmd/gofmt.

Note that cmd/gofmt wasn't in the table,
but it feels like it should be there,
as it's pretty different from the other cmd/... tools.

Change-Id: Ia3455707105b8ab1fcd35799b592de85a9313655
Reviewed-on: https://go-review.googlesource.com/c/build/+/350550
Trust: Daniel Martí <mvdan@mvdan.cc>
Reviewed-by: Robert Griesemer <gri@golang.org>