maintner/maintnerd: increase memory request and limit
maintnerd is restarting frequently, with no notable events in Kubernetes
or anything meaningful in the logs. It uses right around 4GiB of ram, so
it could be running out of memory. Hope this helps.

This also changes to the Recreate strategy, as the default RollingUpdate
will not allow the volume to be mounted in a new container before the
old container goes away.

Change-Id: Ic5e2fc7f8b06d28c83d35a49c32858e34b79187c
Reviewed-on: https://go-review.googlesource.com/c/build/+/356931
Trust: Alexander Rakoczy <alex@golang.org>
Run-TryBot: Alexander Rakoczy <alex@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>