dashboard, cmd/coordinator: support multiple builder owners
Prefer displaying GitHub usernames for convenience of pinging in GitHub
issues, since all current builder owners have them. Keep it possible to
use an email or something else if no GitHub account exists.

Update the owner of the Corellium builders to be individual GitHub
account, so that it can be pinged in an issue. The GitHub organization
is already mentioned in the builder notes.

Also switch to html/template (instead of text/template) for rendering
the builders page.

Fixes golang/go#49596.

Change-Id: I89a96eb8f7a3057bc1f3aaee16abc9c776c45ee7
Reviewed-on: https://go-review.googlesource.com/c/build/+/363983
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>