cmd/release: update upload URL to go.dev
This change also fixes up a bunch of comments and logged messages.

Note that this is necessary because although the golang.org URL leads to
a redirect, it's a 301, which causes the POST to turn into a GET,
resulting in a 405 bad method error.

Change-Id: Ib8521a63b612f293491669b141d068929697ecaf
Reviewed-on: https://go-review.googlesource.com/c/build/+/368894
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Carlos Amedee <carlos@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Trust: Michael Knyszek <mknyszek@google.com>