vcs-test/README.md: warn about Google Cloud Storage default caching
I don't know when the default Cache-Control policy was added, but it
makes it very difficult to fix mistakes in zipfiles uploaded to
vcs-test. Fortunately, the docs for 'gsutil setmeta' give a recipe for
how to fix it, if the user remembers to do so.

For golang/go#35671

Change-Id: I86d365dfae127f3a5230ef4a3c772b17db261096
Reviewed-on: https://go-review.googlesource.com/c/build/+/382594
Trust: Bryan Mills <bcmills@google.com>
Run-TryBot: Bryan Mills <bcmills@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>