cmd/coordinator: preserve repo and subrepo commit times and branches
Benchmarking code needs more information on the commit time for the main
repo revision. However, it seems silly to *just* plumb that. So, let's
plumb down both repo and subrepo commit times and add a helper to
determine which one to use at the point of reference.

At *that* point, it seems silly to only do this for commit times and not
for the branch. Let's err on plumbing additional details for the moment
because they're sure to come in handy for identifying benchmark run
differences down the line. This change does the same thing for the
branch.

Change-Id: I38e97e25b271cbd36d7e9bb2f765211c8555cb7d
Reviewed-on: https://go-review.googlesource.com/c/build/+/393640
Trust: Michael Knyszek <mknyszek@google.com>
Run-TryBot: Michael Knyszek <mknyszek@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Alex Rakoczy <alex@golang.org>