cmd/cl: delete
From information available to me, this command isn't actively used now.
It was created in 2015 to help find CLs that need attention, but by
now there are newer processes (e.g., gopherbot assigning reviewers,
the go.dev/s/needs-review link, the dev.golang.org/reviews page) that
contribute to it being obsolete. Its last meaningful change was in 2016
just to keep it working with Go 1.8.

Having this checked in in x/build implies that it's supported, but that
doesn't appear to be the case. If we did want to keep it supported and
invest more effort into it, it would be better to share more code with
an equivalent web view at https://dev.golang.org/reviews rather than
have two very similar implementations with subtly different logic.

Its code remains available in history if it becomes needed again,
but it's not meeting the bar to stay checked in to x/build.

For golang/go#51867.

Change-Id: I6241bd38d8e828d28a2c31a26bf2d4794c87cf51
Reviewed-on: https://go-review.googlesource.com/c/build/+/394518
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Benny Siegert <bsiegert@gmail.com>