internal/relui: add GCS and staging support
Plumb through the FS implementations we can use to store scratch files
and stage binaries for signing. Use it to do both, replacing the prior
WorkingFiles interface.

Includes some enhancements to the OS fs implementation, reflecting the
fact that it's more of a test fake for the GCS fs than standalone. In
particular, directories are implicitly created.

Introduce an artifact type, used to track the data we'll eventually need
to register it with the web server. This changes slightly in the next CL.

For golang/go#51797.

Change-Id: Iebf53eb43587f5311e428c319d2f3b44aea9fa97
Reviewed-on: https://go-review.googlesource.com/c/build/+/403056
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>