internal/relui: add a step to wait for signing
For the moment, signing will be run out-of-band with the release
process. Add a step that waits for it to complete.

Because our signing process is rough around the edges there are a lot of
subtleties to deal with: we don't produce signatures for zip files,
files only appear in /signed/ if they were modified by the signing
process, etc. Most inconveniently, the .pkg files are produced by the
signing process, which doesn't fit the model I was going for and
requires them to be injected at signing time.

For golang/go#51797.

Change-Id: I6d0b5ffe8b75f4a0d66de3f059d166c843cd0209
Reviewed-on: https://go-review.googlesource.com/c/build/+/410014
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>