internal/workflow: add value-less dependencies
While building the release workflows, I'm frequently needing to express
a dependency on a function succeeding, not a value it produces. For
example, we need to wait for tests to pass, or check that there are no
release blockers. So far I've been expressing those dependencies as
empty strings, but it feels silly: the producing function has to return
"" everywhere, and the consuming function has know that it needs to wait
for something and accept a dummy argument.

Instead, introduce the concept of Actions to the workflow package. An
Action is similar to a Task, but its function returns only an error, and
its result is a Dependency, not a Value. Dependencies can be passed to
any Action or Task to express a dependency between the two, and are
invisible to the dependant function.

I'm not terribly fond of this API, so suggestions for improvement are
welcome.
 - Task and Action are not clearly distinguished. Perhaps Task should be
   renamed Function?
 - I don't like that you can't tell just by looking at a call site which
   of the TaskInputs are arguments that need to match the function's
   arguments. We could try introducing a fluent API or a builder struct.
   I fear either would be too verbose.

For golang/go#51797.

Change-Id: Ibd60b90ce0bf0ae83f7320f6a4e610dde8994d94
Reviewed-on: https://go-review.googlesource.com/c/build/+/410243
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Alex Rakoczy <alex@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>