internal/workflow: generify
Finding out about type errors at runtime is no fun, and precisely the
kind of things generics are intended to prevent. Use them for the
workflow package. It should be impossible to make a type error now.

Due to limitations of Go generics, all methods become functions, many of
which take the Definition as their first argument. Task and Action
functions have to specify their arity.

To compensate for the increase in verbosity, import the package as 'wf' in a
few places. I might fully rename it later.

The implementation doesn't change much. The new Meta interfaces make it
possible to work with heterogenous Parameters and Values. I also took
advantage of knowing task arities to make After return a different type,
so that it's also a compilation error to forget to call it.

The way Params work is a little clunky; having to specify the type both
as a type parameter and as the ParamType struct is redundant. Room for
improvement there.

Change-Id: I1ed957ef441099d00842135243eb2eb9bfda705d
Reviewed-on: https://go-review.googlesource.com/c/build/+/411198
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Jenny Rakoczy <jenny@golang.org>