internal/relui: record resume errors on workflows
When a workflow fails to resume from an error, record the error on the
workflow and mark it as finished. This typically happens when resuming a
task for an un-resumable workflow, either because the definition changed
or is missing.

With this change, the retryTaskHandler can simply redirect, as the user
will see the error on the workflow itself.

Change-Id: I934dc8b24983716820a3b19160052b173d3b519e
Reviewed-on: https://go-review.googlesource.com/c/build/+/412134
Run-TryBot: Alex Rakoczy <alex@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>