dashboard: add "*-newcc" windows builders with updated VMs
Add a set of new "*-newcc" windows 386/amd64 builders that use new
Windows VMs that incorporate updated GCP base images and updated C
compilers. These builders are tagged with a known issue (35006) and
are not included in any trybot sets at the moment. This change is
intended to make it easier to work on the remaining issues blocking
the adoption of the new compilers; once we have fixes checked in we
can get rid of these builders and roll the default windows builders to
the "*v8" images.

Updates golang/go#35006.

Change-Id: I1a7fb5699c20428a846b932e395aeb10ad7adfa9
Reviewed-on: https://go-review.googlesource.com/c/build/+/413415
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Than McIntosh <thanm@google.com>