cmd/relui: join communication tasks in main release workflows
The next step after joining the announcement mail and tweet tasks
is to append them to the end of the main existing release workflows.

This removes the need for the human release coordinator to manually
enter the Version (and, for minor releases, Secondary Version)
parameters, since that information is already automatically computed
as part of those workflows, and can be passed along.

Since the website update is not yet integrated into the release flow,
and as an initial safety check, add another "Wait to Announce" step
that must be approved by the release coordinator, before relui proceeds
to announce a recently published Go release.

This change adds, but doesn't yet enable, this functionality; it's
currently gated on the mergeCommTasksIntoReleaseWorkflows constant.

For golang/go#53537.

Change-Id: Id9046ae3de52da5ba49e4bcecda2ca427fd0e51c
Reviewed-on: https://go-review.googlesource.com/c/build/+/414035
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>