dashboard: make windows race newcc builder actually run race test
A builder is considered as a race builder if its name ends with
-race (see the IsRace function). The newcc builder is named
-race-newcc, which isn't recognized as a race builder. Rename it
to -newcc-race to make it so.

(We could change the IsRace logic. But it is fine and simpler to
just rename the builder, especially as it may be temporary.)

Change-Id: I2a7400855e21b3af4dac6934d65a052be00391fe
Reviewed-on: https://go-review.googlesource.com/c/build/+/415114
Run-TryBot: Cherry Mui <cherryyz@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>