internal/relui: provide full task name to approval task
The approval task needs the task name in order to be able to query the
approval state of itself. For minor releases, the name is prefixed by a
string that is opaque to the running task.

For a long term fix, we could provide the task's full name in the
TaskContext.

Change-Id: Ica2cf4d01517780300e6fade1821265897c38395
Reviewed-on: https://go-review.googlesource.com/c/build/+/417214
Run-TryBot: Jenny Rakoczy <jenny@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>