internal/relui: persist retry-related fields
Store the retry count in the database and test that it all works.

I removed the definition argument from start; we can read it just as
well from the holder, and it makes it harder to forget to register it.

For golang/go#53886.

Change-Id: I4c3d48a96bae87c31629f76067e75ccd524381b0
Reviewed-on: https://go-review.googlesource.com/c/build/+/417590
Reviewed-by: Jenny Rakoczy <jenny@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>