internal/relui: choose major versions dynamically
For convenience, I originally hardcoded Go1.17, 18, and 19 as the live
major Go versions. We're about to release 1.19, so that's not so
convenient any more.

Extract a function to read the live major version, and use it to
choose what workflows to create. While I'm here, add workflows to do
a single minor release in case we need to do that at some point.

For golang/go#51797.

Change-Id: I1e73e42b6c78df7f3e8deb0a191893ff3c887297
Reviewed-on: https://go-review.googlesource.com/c/build/+/418789
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>