buildlet, debugnewvm: fixes for apparent GCE API changes
GCE is returning a URL like
	https://blahblah/projects/symbolic-datum-552/zones/us-central1-a
in inst.Zone, not just us-central1-a.

This is making the code that expects a zone break, but it also gives us
a way to find the project, which we need in a gcloud invocation in case
the default is not the right one. So update the code to use the new
inst.Zone to its fullest potential.

Also adjust the 'probing buildlet' loop to be just a little less aggressive.
It can take a while before it succeeds, and every failed probe prints a large
Python stack trace, making once per second a bit too much.

Change-Id: I4512aa5498cef34ca43b9c50fc47efdbeca15c60
Reviewed-on: https://go-review.googlesource.com/c/build/+/419076
Run-TryBot: Russ Cox <rsc@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>