internal/relui: mail DL CLs after release approved
We put the DL CL early to avoid having to wait for proxy.golang.org, but
now that we're prebuilding, it doesn't make so much sense: either it
hangs around for days or we submit it way early. Move it to the rest of
the publishing steps.

For golang/go#53799.

Change-Id: I12df3019c8278d7c72fe8dc308e209ff3bf5c131
Reviewed-on: https://go-review.googlesource.com/c/build/+/420543
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Jenny Rakoczy <jenny@golang.org>