internal/task: add greenness checking to TagXReposTasks
Before we tag an x/ repo, we want to make sure it's green across all
first-class builders and release branches. That's a surprisingly
complicated process.

First, we need to know which release branches are live; we can get that
from the front page of the build dashboard. Then we need the results for
the individual subrepo. Then we need to decide which builders need to
pass for a CL to be considered green. *Then* we need to know which
commits are on which release branch, which the coordinator knows in some
parts of itself but forgets by the time it serves the dashboard.
Finally, we can combine all that data to find the first CL that passed
after the commit we'd like to tag. It doesn't need to be the exact
commit, in case the build happened to be broken on a non-trybot at that
time.

Given how complex all that logic is, the test is unfortunately also kind
of a mess, sorry.

For golang/go#48523.

Change-Id: I789d83c41383df291e64d466a77c86182be15537
Reviewed-on: https://go-review.googlesource.com/c/build/+/435501
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>